Adds test_distributive.
Only created when Cyclops is available. Also adds the beginning of a
psi4 helper to load data in. tensor::helpers::psi4::load_matrix is
capable of loading a matrix from disk (on master process) no safety
checks are performed. load_matrix is also capable of slicing the data
to distribute it.

From the test directory:

mpirun -n 4 test_distributed

should work and report “norm of S is 2.932177”. Printing a distributed
tensor inside an mpirun does not necessary work..needs to be debugged.