Adds iterate and citerate to Cyclops interface.
iterate implies intent to modify tensor data and thus requires a call
to Cyclops to inform it of the change at the end of the iterations. A
const version named citerate implies that the data will not change and
Cyclops is not informed. Constness is enforced by a const reference.