Adds printn, max, min, and barrier functions.
printn: Each node is allowed to print to the screen with its rank at
the beginning. Each node must call the function.
max: Finds the maximum value in the tensor. Implemented for kCore and
kDistributed.
min: Finds the minimum value in the tensor. Implemented for kCore and
kDistributed.
barrier: When compiled with MPI is equivalent to MPI_Barrier otherwise
it is a nop.